requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.6
django-cors-headers==2.1.0
django-easy-pdf==0.1.1
django-oauth-toolkit==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.7.0
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
jsonfield==2.0.2
oauth2-provider==0.0
oauthlib==2.0.4
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.1
PyPDF2==1.26.0
python-magic==0.4.13
pytz==2017.2
reportlab==3.4.0
requests==2.18.4
six==1.11.0
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.1
wkhtmltopdf==0.2
xhtml2pdf==0.2b1
